+++ philips_hue_v2/bridge.py
@@ -10,8 +10,8 @@ from philips_hue_v2.authentication import get_access_token
 class HueBridge(BaseModel):
     """Class representing a Philips Hue bridge."""
 
-    client_key: str | None = None
-    user_name: str | None = None
+    client_key: str
+    user_name: str
     ip_address: str
     path: Path = Path("bridge.json")
 
@@ -24,10 +24,14 @@ def load_bridge_from_file(filepath: Path) -> HueBridge:
 
 
 def get_access_token_from_bridge(
-    bridge: HueBridge, app_name: str, instance_name: str
-) -> HueBridge:
+    ip_address: str, app_name: str, instance_name: str
+) -> None:
+    """Wrapper function for getting access token from a bridge.
+
+    It outputs the result to the console.
+    """
     result = get_access_token(
-        ip_address=bridge.ip_address,
+        ip_address=ip_address,
         app_name=app_name,
         instance_name=instance_name,
     )
@@ -36,9 +40,5 @@ def get_access_token_from_bridge(
         logger.info(
             f"Successfully got data from HueBridge!\n{json.dumps(data, indent=4)}"
         )
-        bridge.client_key = data["clientkey"]
-        bridge.user_name = data["username"]
-        return bridge
-
+        return
     logger.error(result.unwrap_err())
-    return bridge
